Move MultilineMiddleTruncate into components dir
