+++ packages/studio-base/src/components/DataSourceSidebar/DataSourceInfo.tsx
@@ -13,7 +13,7 @@ import Stack from "@foxglove/studio-base/components/Stack";
 import Timestamp from "@foxglove/studio-base/components/Timestamp";
 import { subtractTimes } from "@foxglove/studio-base/players/UserNodePlayer/nodeTransformerWorker/typescript/userUtils/time";
 
-import { MultilineMiddleTruncate } from "./MultilineMiddleTruncate";
+import { MultilineMiddleTruncate } from "../MultilineMiddleTruncate";
 
 const selectStartTime = (ctx: MessagePipelineContext) => ctx.playerState.activeData?.startTime;
 const selectEndTime = (ctx: MessagePipelineContext) => ctx.playerState.activeData?.endTime;
