+++ packages/studio-base/src/components/MultilineMiddleTruncate.tsx
